pulled filter hooks into filters file in utils, pulled checkbox into checkbox file in utils, hooks-based affix and root table files are cleaned up and renamed for clarity
